backend/processing_service/requirements.txt
boto3>=1.26.0
Pillow>=9.5.0
